+++ src/components/common/FoodBtn.tsx
@@ -12,6 +12,7 @@ export default function FoodBtn({
   return (
     <button
       onClick={onClick}
+      type="button"
       className={`lg:px-[0.625rem] lg:py-[0.2813rem] lg:text-sm text-dark-green border border-dark-green rounded-full whitespace-nowrap
         ${selected ? 'bg-orange text-white border-orange' : ''}
       `}
